Ughh....added delegate params.
Experimental attempt to re-add the ability to add params to the delegate functions. I don't think this is an acceptable approach. Although it kind of works, it introduces even more edge-cases. The problem is that the precedence of DI information set through the other functions, ->make($foo, $extraParams) and now ->delegate($foo, $extraParams) is fundamentally a mess. There will be no precedence that is correct for all usages.

Because of that, it's probably better to leave it out and instead tell people to do the wiring up by hand, or just use Auryn as a service-locator where appropriate.